pyproject.toml
[project]
name = "q2api"
version = "0.1.0"
description = ""
authors = [
    {name = "q2api", email = "admin@example.com"},
]
dependencies = [
    "fastapi==0.121.3",
    "uvicorn[standard]==0.38.0",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.3",
    "httpx==0.28.1",
    "websockets==15.0.1",
    "aiomysql==0.3.2",
    "asyncpg==0.30.0",
    "PyMySQL==1.1.2",
    "aiosqlite==0.20.0",
    "aiofiles>=24.1.0",
    "limits==5.6.0",
    "tiktoken==0.12.0",
    "requests==2.32.5",
    "slowapi==0.1.9",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.0.0",
    "cryptography>=42.0.0",
    "psutil>=5.9.0",
    "PyJWT>=2.8.0",
    "sniffio>=1.3.0",
    "sortedcontainers>=2.4.0",
    "bcrypt>=4.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
]

[tool.hatch.build.targets.wheel]
include = [
    "app.py",
    "src/",
    "frontend/",
    "templates/",
]
